+++ package.json
@@ -98,6 +98,7 @@
     "react-csv": "^2.0.3",
     "react-dom": "^16.13.1",
     "react-dropzone": "^11.3.1",
+    "react-error-boundary": "^3.1.4",
     "react-keyboard-event-handler": "^1.5.4",
     "react-mosaic-component": "^5.0.0",
     "react-redux": "^7.2.0",
